[HFIP_2024] Add mpassit tasks (#22)
Adds MPASSIT as a new component in the build system, and as a task in the workflow. Makes some other defaults changes to accommodate this new component (modifications to the configuration of the forecast, for example). Also adds the parm files to this repo so we aren't reliant on other locations for text files.